feat(k8s): apply container service annotations to Pod templates as well
When using the Kubernetes provider, the annotations set on
`services[].annotations` in container modules are now applied to both
Service and Pod resources. You can generally specify the annotations
intended for both Pods or Services here, and the ones that don't apply
on either side will be ignored (i.e. if you put a Service annotation
here, it'll also appear on Pod specs but will be safely ignored there,
and vice versa).