feat(template): add template helper functions
This adds a collection of helper functions to our template syntax.

Initially we've added a handful of specific helpers that users have
requested, but we'll be very happy to add more since that generally is
very simple to do now that the basic structure and syntax support is in
place.

See the added docs for the usage instructions and the helpers available
in the first iteration.

_Reviewers, please don't hesitate to suggest more functions to include
in the first iteration!_