feat(config): add `this` context when resolving modules
This adds `${this.name}`, `${this.buildPath}` and `${this.path}` to the
template context when resolving modules.

Because someone might ask, `${this.version}` is not included because the
module version is determined after resolving the configuration.

I'm adding this both as a convenience and as a part of a different
project, where I'll add further keys to the `this.*` context.