feat(commands): skip-dependants opt for test cmd
The current behavior of the test command when called with a list of
module names is to run all tests in those modules, and all tests in
modules with dependencies on those modules.

This is not the desired behavior in all cases, so here we add a
`--skip-dependants` option to the test command. When used, only tests in
the requested modules will be run.