feat(k8s): resolve template strings in kubernetes module manifest files
Previously, you could only use template strings when inlining
manifests using the `manifests` field. Now, template strings are also
resolved in the files referenced in the `files` key.

Internally, we now make a `ctx.resolveTemplateStrings()` for all
plugin handlers, so this will be easy to do in other scenarios as well.