improvement(cloud): better API errors
We now throw an error when a null client auth token / jwt is received
from the backend during login.

Previously, we were only logging an error, without indicating that the
command had failed. The error message was also a bit cryptic.

Also, when workflow run registration fails due to CLI/backend API
incompatibility (i.e. if the shape of the registration request isn't
what the API expects), we now throw a better error, and include the
failed request payload in the error detail.