chore: add vaccine watcher command
This command checks for openings at Berlin vaccination centers, using
the API from Impstoff.Link. Obviously this is only useful for people in
Berlin, and only during the vaccination efforts. A big shout out and
thank you to the people behind Impstoff.Link for making this possible.
And of course to everyone working towards getting Berliners vaccinated.