feat(core): local targets for build dependencies
Added an optional `copyToSourceDir` flag to
`build.dependencies[].copy[]`.

When used, build dependencies are copied to the module (source)
directory of the dependant module instead of the build directory.

This is useful for many flows, especially when using the `exec` module
for custom build flows, and has been indirectly requested by many of our
users.