fix(k8s): omit probes in runner pod spec
We now no longer include liveness or readiness probes on the pod spec
used when running e.g. tests or tasks for `kubernetes` or `helm`
modules.

In certain situations, these probes could result in test pods being
evicted before the test suite was able to finish. This is fixed here.