feat(core): stream build statuses
We now stream build status events, similarly to how we've been streaming
service, test and task status events.

We've also added a `running` state to the `RunState` type, and we
include a `building` state in the new `BuildState` type to represent
builds/tests/tasks that are in progress.

Additionally, we now emit a `serviceStatus` event with the `deploying`
state in the `deployService` action just before the deploy handler is
called.

Finally, we've also added `moduleVersion`, test/task/service config
versions and an optional `actionUid` to the corresponding status events.