fix(core): don't omit dev mode services from deployment dependencies
This was done a little carelessly. There's a different behavior for
dev mode, compared to hot reload, in that we don't expressly signal
whether a service supports dev mode. So when running the dev command
or `deploy --dev=*` this would in effect completely mess up the
dependency order of actions.