feat(helm): return ingresses on helm getServiceStatus()
We can't see the ingresses created by the `helm` provider and this
behavior really decreases the quality of our development
workflow. Having the list of deployed ingresses in the CLI output or
the dashboard is very convenient.

I am not a typescript guy and I didn't write tests, but if you approve
the direction of this PR I can work on that.

We are planning to use Helm charts to deploy our own code since we
want to have those charts as part of the development flow outcome.