improvement(core): don't watch dev-enabled modules
This is a performance optimization that lowers resource use when using
dev mode. We now don't watch files inside the module roots of modules
having one or more services deployed with dev mode, except for the
modules' Garden config paths.

The rationale here is that FS watches can be resource intensive,
especially when there's a lot of churn (and especially on Linux). Since
Mutagen takes care of the dev-mode sync (and doesn't go through the
framework-native watch logic), this is usually unwanted overhead.