fix(core): fix handling of empty option values
To ensure that `command.params` behaves intuitively in template strings,
we don't add option keys with `null`/`undefined` values to the command
info object (which gets passed to `command.params` in template strings).

For example, we don't want `${command.params contains 'dev-mode'}` to
be `true` when running `garden deploy` unless the `--dev` flag was
actually passed (since the user would expect the option value to be
an array if present).