improvement(cloud): more robust log streaming
We now reuse the machinery from the new Kubernetes logs follower for
gathering real-time log output from `runAndCopy` when run without
artifacts.

This should ensure that all log lines are captured, and makes the log
gathering process more robust when the connection to the pod fails and
has to be reestablished.

When the connection has to be reestablished, we replay past log lines
and deduplicate them against a fixed-size buffer of log entries for the
pod + container in question to ensure that no lines are lost, while
not emitting duplicate log lines.