feat(core): allow sparse arrays for more fields
We now allow sparse arrays for many more config fields, such as
`services[].args` (and `command`) and in workflow step commands.

This can now be done in most places where arrays of strings are provided
in configuration.

For example, this will now work in a workflow step command (where the
step command will be run without additional arguments if
`var.serviceNames` is null):
```
[deploy, ${var.serviceNames ? var.serviceNames : null}]
```
This makes it easier to implement highly dynamic configuration and
reduce repetition in certain cases.