feat(k8s): add two-way-resolved-reverse option
This should be the last Mutagen sync mode we'll have to add.

When using a resolved two-way sync, it's useful to be able to decide
which endpoint is used as the primary. With this sync mode, the user now
has every combination of options when using the resolved one/two-way
sync modes.