feat(core): render module command
Added a new command: `garden render module`.

This command resolves all template strings in the module, and
outputs the full module configuration (which includes the resolved
module version).

This is useful for debugging/writing template strings for modules, and
for reasoning about which files/variables factor into the  module's
version calculation.