feat(cli): custom commands
This adds the ability to define custom commands as part of any Garden
project. See the added guide for details.

When reviewing, please try making your own and see if you bump into
any issues or get confused!

I'm flagging this as experimental just for the moment, since this sort
of thing invites a fair number of edge cases and use cases we didn't
consider up front, and getting real-world feedback will be important.

Some additional notes on the PR:

As part of the process I started implementing a "spread" option
for positional arguments, but went another way in the implementation,
so finishing that will wait but I left those initial changes in.

I had to remove the preemptive validation for workflow commands, and
along the way I allowed more commands to run in workflows by making the
criteria whether they're persistent or not.