fix(core): improve project config validation
We now validate that `project.environments` is an array and throw an
informative error if it isn't.

Previously, if `project.environments` was mistakenly formatted as a
map instead of an array, a cryptic type error would be printed ("Cannot
read property name of undefined"). We now clearly indicate the problem
to the user.

Additionally, the `joiSparseArray` schema can now properly handle
non-array inputs without failing cryptically.