fix(workflows): fix provider resolution issue
This fixes a regression introduced in the last release where providers
would be resolved ahead of any script steps in a workflow, which caused
problems when using a script to e.g. prepare auth credentials that were
subsequently used when resolving providers.

We now don't emit a stack graph event for workflows that don't have a
command step that resolves the config graph.