chore: use 'Garden Cloud' instead of 'Garden Enterprise'
Garden Cloud and Garden Enterprise are distributions of the same
application. To keep things simple, we default to referring to the
product as Garden Cloud.

This PR removes old references to Garden Enterprise from text and
ensures cloud specific commands can be called both as `garden cloud
<args>` and `garden enterprise <args>` for backwards compatibility.

Where we can read the string dynamically and use Garden Enterprise
or Garden Cloud as applicable.