fix(core): properly handle sparse build deps
Previously, adding an optional build dependency via a template string
that sometimes resolves to null would throw an error when the string
resolved to null.

The build dependencies array was always intended to be a sparse array,
but due to the special handling we have to support both the shorthand
(module name only) and the full (a map with the module name and other
fields) formats, we needed to add a bit of extra logic to account for
this use case.