feat(core): debug logfiles
We now output all debug-level and silly-level logs into logfiles in the
project's `.garden/logs` directory.

Debug-level logs are logged to a debug logfile using standard text
formatting.

Silly-level logs are logged to a JSONL-formatted logfile (where each
line is a JSON string). These can later be used to provide dynamic
rendering of command logs via the CLI (e.g. allowing the user to select
the log level, and then filtering the JSON entries before rendering them
to the terminal).

The names of these logfiles include the command name, log level and the
time the command was executed (as an ISO timestamp).

Logfiles older than 7 days are also automatically cleaned up by this
logic.

This will be useful for gathering debug information when something
goes wrong in a Garden command that was run at a lower (less verbose)
log level.