chore(k8s): update k8s-sync Dockerfile for Mutagen v0.13.0
Mutagen no longer builds or ships individual agent binary archives, so
we now need to extract the agent binary from the agent bundle in the
full release archive.

We can also switch to official Mutagen releases for the agent binary
because the Garden fork doesn't make any modifications to the agent.