improve(cli): faster initialization
I did a bunch of profiling to work out where time was spent on basic
initialization, before actually doing anything meaningful. The changes
combined shave about 50% off the time to run e.g. `garden help` in my
testing.

The only semantic change is the removal of the fullscreen writer, which
was never really made usable and imported some heavy libraries.

Beyond that, it's a lot of changes to lazy-load libraries when needed,
and in particular to load plugin code later, and avoid loading sqlite
until actually necessary.

There was also a seemingly unnecessary call to clean up the Mutagen
daemon directly from the Garden.close() method. I'm not exactly sure
why that was added, but if it's actually necessary we should find
another way to do it, instead of bleeding plugin code to the core Garden
class.