feat(core): support persistent local processes
This commit adds proper support for persistent local processes.

It introduces the 'persistent' flag to the service spec of the local exec
module type.

When the 'persistent' flag is set to true and Garden is in watch mode,
Garden does not wait for the local process to exit before considering
the service deployed.

Users can however set a status command that checks whether the process is
actually 'ready'.

It also adds support for streaming logs from local processes. This is
achieved by streaming the output of local procs to a file and then
tailing the file when the logs command is run. Turns out there's no
proper (i.e. efficient) npm library for this so the bulk of the logic is
actually around file tailing.

I wouldn't be surprised if we run into edge cases or platform specific
gotchas here but they should be easy to fix when they come up.