improvement(core): evaluate null|false as undefined for dockerfile field
We want to better handle optional strings in general at the schema
validation level by  mapping null and false to undefined.
However that's a bigger change that requires proper testing and
documentation.

We'll do that in a follow up PR but just getting this small fix out now
to unblock some users.