fix(core): attempt to fix unhandled promise rejection from got
Users have been reporting the following error when using Garden:

error: unhandledRejection: The `onCancel` handler was attached after
the promise settled.

I haven't been able to reproduce it but apparently it's an issue with
got. See: https://github.com/sindresorhus/got/issues/1489

This commit _attempts_ to fix it by implementing the workaround
suggested here:

https://github.com/sindresorhus/got/issues/1489#issuecomment-805485731

Some users have reported that it fixes the issue, others that it
doesn't. Since I haven't been able to repro myself this is sort of a
best effort thing. Hopefully it does the trick.