fix(core): resolve runtime values in test configs
Before this fix, `${runtime.*}` values were not resolved in test configs
before calling the plugin handlers for `getTestStatus` and `testModule`.

This was fixed by applying a similar runtime resolution flow as is used
for service and task actions to test actions.

For consistency, a few interfaces and types were added to correspond
with how these things are laid out for service and task handlers.