feat(core): opt for deleting services in dep order
Added a `--dependants-first` option to the `delete service` and
`delete environment` commands.

When used, services will be deleted in reverse dependency order. That
is, service dependants will be deleted before their dependencies.

Also added a `--with-dependants` option to the `delete service` command,
which deletes any running dependants of the requested service/services.