improvement: support only single dotIgnoreFile (#3069)
BREAKING CHANGE:

The `dotIgnoreFiles` field in project configs has been deprecated in 0.13 in favor of the `dotIgnoreFile` field, and as of 0.13 only one filename is allowed here.

If a single filename is specified in the old `dotIgnoreFiles` field, the conversion is done automatically. If multiple filenames are provided, an error will be thrown.