feat: allow module overlap if one or both of them are disabled (#3222)
* feat: allow module overlap if one or both of them are disabled

This fixes module validation to allow for modules to overlap, if
one of them is disabled. Example use case: one environment deploys
to kubernetes, the other environment deploys to a serverless env.

This also changes the error message: The goal was to make clear that
there are two ways to get rid of the error if the overlap is intentional.