chore: disable edge releases on master
We've started getting complaints from users that they're being spammed
with release notifications for every edge release.

For the time being, we'll disable edge releasees on merges to master. We
may re-enable them once we've had a closer look and found a way to do
this in a less noisy way.