docs: use Discord instead of Slack for community links
We're migrating our community from the private Kubernetes Slack
workspace to a Discord server that we own.

This commit updates links in docs and code accordingly.

See also this Slack announcement for details:

-------------------------------------------------

In an effort to streamline our centers of community action,
we at garden.io have decided to sunset our Discourse and Slack
channels with a one month notice from today.

Why? Two reasons:

1. It's time to leave the nest.

Our Slack channel is under the auspices of the Kubernetes moderation
team and we couldn't be more grateful for their efforts to make
the Slack the warm and welcoming place it is. As we grow and
add open source users it's important we have all the tools
we need to (kindly) moderate our community and Discord simply has
a better suite of tools for that task.

2. The posts on Discourse are all GitHub issues.

Almost universally, the posts on Discourse have already been
posted to our GitHub or are GitHub issues in disguise :disguised_face:.
In the majority of those, posts already mirrored to GitHub are to add
detail or to draw another lane of attention. That's on us for not circling
back to GitHub issues. One effect of sunsetting the Discourse and
Slack channels is it gives more time back to our team to ensure
those issues are getting the attention they need.

It bears mentioning that we have no community moderators (yet) on
our team which means that everyone gets fastidious attention from
our truly wondrous OSS devs and SREs.

To recap: we will be closing the Discourse and Slack channels November 14th.

-------------------------------------------------