chore(cloud): don't emit client requests as events
Previously, we were emitting build, test, deploy and task requests on
the event bus and handling them in `processModules`.

This was unnecessarily complicated, and resulted in requests from Cloud
being emitted back to it as events (which wasn't the desired behavior
for our event logging).

Now, we handle these as requests coming in (instead of emitting them
as events). The old Cloud UI event payloads have been repurposed as
types for these requests' payloads.