fix(cloud): fix duplicate event emission
Before this fix, we were sending each event twice to websocket
subscribers in live Cloud sessions.

This is because the `ignoreHost` logic in `DashboardEventStream`
expected a `serverHost` without an auth token appended, which was
broken when the dashboard auth tokens were introduced a while
back. The fix was to simply compare the base server URLs.