fix(k8s): compareDeployedResources should not ignore manifest hash
compareDeployedResources actually just skipped isSubset check in case
the manifest hash was present.

Now it actually returns "outdated" when manifest hash differs.

This is necessary if the manifest removes annotations that are
present in the deployed manifest.

E.g.:

1. upsert annotations: {"some-annotation": "present"}
2. upsert annotations: {}

before this change, compareDeployedResources just returned "ready"
even though the deployedManifest was actually outdated (the annotation
needed to be removed)

This commit fixes that