fix(logger): fix issues with spinner and duplicated entries
Before this fix, the spinner would be out of place (top left corner) if
the entry it belongs to is outside of the terminal viewport.

Furthermore, updating entries outside of the terminal viewport would
cause some log lines to be rendered twice. This was only visible when
scrolling back.

The commit also adds an env var which allows users to fallback to the
previous render logic. Terminals are fickle beings and it feels good to
have an escape hatch if this breaks, say, tmux on WSL.

We can remove the env var later if there are no complaints.