fix(framework): idempotent config store deletes
Previously the ConfigStore class would throw if attempting to
delete an entry with a key that doesn't exist.

So e.g. running delete twice on the same key would fail the second time.

I'm 99.9% sure this that is just a mistake that was made years ago and
hasn't surfaced before.

This fix is needed to handle a test error on a different branch.