fix: send a periodic ping over the websocket for exec
Currently it can happen that the websocket connection is timed out somewhere in the network, and no error event is triggered.
Garden believes something is still executing, and never receives the information about completion or failure.
This sends a periodic ping so that hopefully the connection stays alive.

As an extra step we could implement the full timeout logic in https://github.com/websockets/ws#how-to-detect-and-close-broken-connections