fix: support local builds in microk8s in multipass
When microk8s runs within a multipass VM, we need to explicitly transfer any files into the VM before having access to them.
This broke builds with local docker on macOS.

Since there is no simple way of determining whether microk8s runs within multipass,
we instead ask multipass to list all its VMs and see if the default microk8s machine name is included.
In that case we assume that we are running in multipass and transfer the file over.