fix(k8s): fixed error handling in image builders
This patches #3388. The method `PodRunner.runAndWait()` is used in a few places. Now it throws an error on non-zero exit code only when it's called from `runWithoutArtifacts(...)`, where the error can be handled correctly. In the rest places, it returns unsuccessful result w/o throwing any errors, just like it was before #3388.

The whole error handling and propagation flow can be revisited in #3254.