fix(logger): always show sections with basic logger
Before this fix, basic logger entries would regularly be 'orphaned'
because the parent section wasn't printed. This made it very hard to
tell what section a given log entry would belong to.

E.g. when the log level is set to verbose, it's hard to tell if build
logs belong to one module or the other.

Furthermore, the commit also:

- Prints the name of the command when rendering output streams.
- Ensures the header isn't rendered twice with the basic logger.
- Removes superfluous arrows from some log strings which don't render
  nicely with the basic logger.
- Prints the log level if using the basic logger and the level is higher
  than 'info'.
- Skip left padding for basic logger.