improvement(config): better debug info for certain schema errors
This adds a createSchema() helper and a name key to supported schema metadata,
which is used to add error detail, and also slightly improve startup performance
by caching created schemas.

It would be good to follow this up by at least adding the name metadata to more
schemas, but ideally to port schemas gradually to use the new helper.