fix: re-initialize providers changing environments (#3481)
* fix: re-initialize providers changing environments

When users change environments, we need to re-initialize providers.

Let's consider the following example:

User has a provider config like so:
```
providers:
  - name: exec
    initScript: echo ${environment.name} > someFile
```

If user runs a garden command with --env foo, and then with --env bar,
and then again with --env foo we would expect `someFile` to contain `foo`. But
the file actually contains bar.

Reason for this is that garden skipped executing the initScript in the foo env,
because it has already been executed. But the cache logic is not aware that the
last command that ran was in the bar env.

This commit fixes that problem.

* improvement: remove unnecessary check for env name

Reasoning is that when the env name changes and that is relevant,
the provider config changes too (assuming that we do not implicitly
pass additional information like the environment name using env
variables)

* tests: add an integration test

Co-authored-by: srihas-g <srihas-g@users.noreply.github.com>
Co-authored-by: Walther <Walther@users.noreply.github.com>