refactor: rename `run task` command to `run`
This completes the removal of `garden run` as a category of commands. Instead,
`garden run` is now an analog of other action verb commands (build, deploy,
test).

There is still a bit TODO in terms of making the command more analogous in
implementation to the other action verb commands, which will follow.

The command will print a warning if it appears the user is trying to run one
of the old `run <subcommand>` commands.

BREAKING CHANGE:

The `garden run task` command has been renamed to `garden run`.