improvement: add implementation to "log" event
* fix: make type inference work for PluginEventBroker

Type inference did not work before this commit for event emission
and event handling. Changing PluginEvents to be a dictionary of
function signatures fixes that.

* improvement: add implementation to "log" event

Add an `implementation` field, enabling plugins to send a short string
with every log message that helps identify the underlying code or tool
that generated this log message.