fix(core): correctly ignore watches in dev mode
This fixes a regression in 0.13 where an `actionSourcesChanged` event
would get fired when sources are changd for a Deploy that's deployed
with dev mode enabled. This caused several problems, but is now fixed.

The problem lay in the way we were calculating the paths to skip/ignore,
which are passed to the `Watcher` class.