feat(kubernetes): add sync helper commands
This commit adds Kubernetes plugin commands for getting the Mutagen sync
status, pausing syncs, and resuming syncs.

I also refactored the Mutagen code to use a class which makes it a
little easier to manage some of the daemon process life cycle logic.

The class is currently a singleton but should be refactored after v0.13
into a "normal" class that we can attach to the `garden` instance.

Finally, the new MutagenDaemon class also logs more information about
what Mutagen is doing under the hood. This is e.g. useful when
Mutagen is syncing large change sets because without the added
logs it looks like nothing is happening.