fix(events): update log event timestamp type be string or number
This essentially reverts a change that was made in
https://github.com/garden-io/garden/pull/3576
(commit: b52e9b298f7c59b8210a77edc4c451301daa76e3).

The type has always been 'number' but we were in fact sending
strings to Garden Cloud which is the payload it expects.

Garden Cloud (version >= 1.360) now supports numbers but we need to
continue sending strings for backwards compatibility.

We can tighten the type to just 'number' once all Cloud instances are
up to date.