refactor: remove old `garden dev` command implementation
For now, the `dev` command implementation is basically that of the `serve`
command, with an added "under construction" message to steer users to the
`deploy --dev` command.

Later we will add interactivity on top, hopefully before the 0.13 GA.

BREAKING CHANGE:

The behavior of the `garden dev` command is being fully replaced. It will no
longer automatically perform all tests and deployments, but instead connect to
the new dashboard, and later feature a command input in the terminal.