fix(events): use ISO format date strings in event payload
Previously the event payload type for dates was a JS Date instance.
When these events are emitted to other consumers such as Cloud the dates
are marshalled to ISO strings.

This commit updates the payload types to strings so that the types
reflect what the consumers actually receive and so that the types can be
re-used across consumers.