fix(logs): ensure invalid dates are handled properly
This bug was introduced when we changed `timestamp.getTime()` to
`timestamp.toISOString()`.

Turns out that the former returns `NaN` if the date instance is invalid
whereas the latter throws. So this had basically been failing silently
until that change was made.

This commit ensures we check for the validity of the date instance.