improvement(logger): remove fancy logger
BREAKING CHANGE: This commit removes the 'fancy' logger.

That is, the default Garden logger that renders spinners and updates log
lines in place.

From now on, the logger previously knowns as the "basic" logger is the
default.

This results in the following breaking changes:

- You can no longer set the logger type to 'fancy' via the `--logger-type`
  flag or the `GARDEN_LOGGER_TYPE` env  var. Note that `"fancy"` was the
  default behaviour so it's very unlikely that it's set anywhere via either
  of the aforementioned options.

Furthermore, setting the logger type to `"basic"` has been deprecated and
will be removed in a future release. Instead the type can be set to
`"default"` (which is kind of a superfluous option).

Furthermore, the logs rendered in the terminal will "look" different
compared to the current default, although that's not technically a
breaking change.

Finally, the shape of the log data sent to Garden Cloud has changed and
will need to be handled specifically in Cloud.

Goodbye, you beautiful disaster.