feat(exec): add explicit `shell: true` option to exec action specs
This is done in part to address a somewhat unexpected default behavior when
exec-ing commands, and to allow a smooth transition from 0.12.x and modules.

Basically setting `spec.shell = true` on any exec action means it will follow
the same behavior as in 0.12. Accordingly, this flag is set when converting
exec modules to actions at runtime.