refactor: rename "dev mode" to just sync or "sync mode"
In the spirit of simplifying terminology and preferring transparent over
opaque (or overloaded) phrasing, we've opted to drop the "dev mode" concept in
favor of simply "sync". This more directly describes the functionality and
should be easier to grok.

Specifically, these are the user-facing changes:
1. Users should now use `--sync` instead of `--dev` in the command
line.
2. The `devMode` configuration fields are now called `sync`.
3. The `devMode.sync` and `devMode.syncs` fields are now simply `sync.paths`.

All these changes are backwards compatible.