feat(pulumi): allow CLI options for plugin cmds
Any CLI options that aren't used by the Pulumi plugin commands
themselves are now passed along to the Pulumi CLI when running plugin
commands.

For example, `--skip-dependencies` won't be passed to the Pulumi CLI
(since it's a Garden-specific option that doesn't apply to the Pulumi
CLI), but the `--same` option would be passed along unchanged to
`pulumi preview`.

For example, in `garden plugins pulumi preview -- some-service --same
--suppress-outputs`, `--same` and `--suppres-outputs` would be passed
as CLI options to `pulumi preview`.