chore(cloud): update action status & log events
Here, we udpate and simplify the action status event payload types used
to communicate with Cloud.

The event payload types have been streamlined and standardized. We now
use `actionState` as the primary state indicator field, which should
reduce the special-casing needed on the backend (since we can interpret
a common sum type for status events from all the main action kinds).

The action states have also been documented thoroughly, which should
help newcomers understand the larger flow more easily.

Soon, we'll want to generate `actionUid`s in the constructors of
`BuildTask`, `DeployTask`, `RunTask` and `TestTask`
and pass them to the corresponding status checking & processing
actions, so that we can track the status check action and the
subsequent processing action (if any) under a single `actionUid`.

We'll be able to utilize this in the Cloud UI e.g.  for flame charts
that capture the full lifecycle of the framework's primary action kinds.