ci: add edge-bonsai tag to image release
For testing edge-bonsai it would be good to build and release container images for use where CI depends on a container image e.g. GitLab CI. Users sourcing the Garden binary as part of a multi-stage Dockerfile strategy should also be pleased with the option to have an edge release available.