improvement(container): redeploy when spec changes
In 0.12, we had wanted to update the `selector.matchLabels` used for
`container` deploys, but doing so would require redeploying those
resources (since `spec.selector.matchLabels` is an immutable field).

In 0.13, we decided to simply automatically redeploy any `container`
deploys if needed (unless the environment in question is a
production environment), since we wanted to use different labels
(e.g. `action` instead of `service`).

Deleting the resources before redeploying them is necessary in this
case, since the `spec.selector` field is immutable and attempting
to update it for running resources will simply result in an error).

We don't automatically do this for production environments, except when
force-deploying (which is a framework-native way for users to redeploy
services with full awareness what they're doing).

We now also use resource-keyed maps of manifests and deployed resources
in `compareDeployedResources`, since the old zip-based approach was a
bit dodgy (since it didn't guarantee that the keys of the manifest and
deployed resource matched).