feat(config): add $if/$else/$then conditional objects
This is analogous to `$forEach` objects, and addresses a common need (that I
happened to run into while working on something else).

From the docs:

---

### If/else conditional objects

You can conditionally set values by specifying an object with `$if`, `$then` and
(optionally) `$else` keys. This can in many cases be clearer and easier to work
with, compared to specifying values within conditional template strings.

Here's an example:

```yaml
spec:
  command:
    $if: ${this.mode == "sync"}
    $then: [npm, run, watch]
    $else: [npm, start]
```

This sets `spec.command` to `[npm, run, watch]` when the action is in sync mode,
otherwise to `[npm, start]`.

You can also skip the `$else` key to default the conditional to _no value_
(i.e. undefined).