fix(core): improve graph result filtering
Fixes an OOM error that would occur when several actions failed
during a command run.

We now mostly select a picked list of fields for including in our
command result, instead of skipping fields (which has historically
led to new fields with unbounded size slipping in, which calls for
repeated fixes around this).