fix(container): fix spec.image in container conversion
When `module.spec.image` is provided, but the value has no tag, we
weren't setting the Build/module version as the default tag/suffix to
the image ID.

In 0.12, this was done by referencing the deployment image ID from the
container build's outputs, which we can't do here. Therefore, we simply
call the relevant helper to shim that behaviour here.