feat(k8s): kubernetes-exec Runs and Tests
Here, we introduce new Kubernetes-based action types for Runs and Test
(both have the action type `kubernetes-exec`).

These run the provided args in an already deployed resource in the
cluster, using \`spec.resource\` to locate the resource.

This is a great fit for inner-loop development, where rebuilding before
retesting is not always desirable (e.g. when running tests in
conjunction with code syncing), and to speed up the spinning up a new
environment by executing Runs in already running services.