fix(k8s): retry exec attempts in PodRunner (#3956)
* fix(k8s): retry exec attempts in PodRunner

Before this change, the Kubernetes API request sent as part of PodRunner
exec calls have not been retried.

This change makes sure that we use the existing Kubernetes API retry
implementation and at the same time it makes sure that HttpErrors are
wrapped properly, for improved error messages.

* fix(k8s): wrap k8s errors for better error messages

This change makes sure that we wrap exec k8s errors like we do for
other Kubernetes API endpoints. This improves the error messages.

* improvement: only retry if the error is recoverable

Only retry on 500, 502, 503 and 429 errors to reduce the risk that the
command is being executed twice.

co-authored-by: tim <tim@garden.io>

---------

Co-authored-by: tim <tim@garden.io>