feat(config): templating for actions and workflows + new RenderTemplate kind
This adds support for templating actions and workflows in ConfigTemplate
configs (formerly ModuleTemplate).

The old way of referencing templates (a Module with `type: templated`) is also
superseded by a new `kind: RenderTemplate` config, with largely the same
semantics, except with an added `configs` field that supports actions and
workflows.

The resolution flow is very similar for actions and workflows, as it was (and
remains for now) for modules.

For now we have left out the option to generate files in the source tree, which
is possible with the now-deprecated `generateFiles` field on modules. We may
later consider making that a native feature on `RenderTemplate` configs as
opposed to on the templated configs themselves, but it may be an unnecessary
feature since that can in most cases be handled by `exec` actions.